+++ src/main/java/com/kcy/fitapet/domain/oauth/dao/OauthRepository.java
@@ -4,10 +4,11 @@ import com.kcy.fitapet.domain.oauth.domain.OauthAccount;
 import com.kcy.fitapet.domain.oauth.type.ProviderType;
 import com.kcy.fitapet.global.common.repository.ExtendedRepository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 public interface OauthRepository extends ExtendedRepository<OauthAccount, Long> {
-    Optional<OauthAccount> findByOauthIdAndProvider(Long oauthId, ProviderType provider);
-    boolean existsByOauthIdAndProvider(Long oauthId, ProviderType provider);
+    Optional<OauthAccount> findByOauthIdAndProvider(BigInteger oauthId, ProviderType provider);
+    boolean existsByOauthIdAndProvider(BigInteger oauthId, ProviderType provider);
     boolean existsByEmail(String email);
 }
